GOB: Split HTML parser classes into a new file (Adibou2/Sciences)
