+++ engines/gob/inter.h
@@ -33,6 +33,7 @@
 
 #include "gob/goblin.h"
 #include "gob/variables.h"
+#include "gob/html_parser.h"
 #include "gob/iniconfig.h"
 #include "gob/databases.h"
 
@@ -791,74 +792,6 @@ protected:
 	void o7_getFreeDiskSpace(OpGobParams &params);
 
 private:
-	static const uint32 kMaxNbrOfHtmlMarks = 20;
-
-	struct HtmlMark {
-		uint32 _field_0;
-		uint32 _field_4;
-		uint32 _field_8;
-		uint32 _field_C;
-		uint32 _pos;
-		uint32 _field_14[5];
-		uint32 _field_28;
-	};
-
-	class HtmlContext {
-	public:
-		HtmlContext(Common::SeekableReadStream *stream);
-		~HtmlContext();
-
-		enum HtmlTagType {
-			kHtmlTagType_None = -1,
-			kHtmlTagType_Error,
-			kHtmlTagType_OutsideTag,
-			kHtmlTagType_Body,
-			kHtmlTagType_Font,
-			kHtmlTagType_Font_Close,
-			kHtmlTagType_Img,
-			kHtmlTagType_A,
-			kHtmlTagType_A_Close,
-			kHtmlTagType_Title,
-			kHtmlTagType_Title_Close,
-			kHtmlTagType_HTML_Close,
-			kHtmlTagType_BR,
-			kHtmlTagType_P,
-			kHtmlTagType_P_Close,
-			kHtmlTagType_U,
-			kHtmlTagType_U_Close,
-			kHtmlTagType_B,
-			kHtmlTagType_B_Close,
-			kHtmlTagType_EM,
-			kHtmlTagType_EM_Close,
-			kHtmlTagType_I,
-			kHtmlTagType_I_Close,
-			kHtmlTagType_SUB,
-			kHtmlTagType_SUB_Close,
-			kHtmlTagType_SUP,
-			kHtmlTagType_SUP_Close,
-		};
-
-		static Common::HashMap<Common::String, HtmlTagType> *_htmlTagsTypesMap;
-		static Common::HashMap<Common::String, char> *_htmlEntitiesMap;
-		static Common::String popStringPrefix(const char **charPtr, char sep);
-		static HtmlTagType getHtmlTagType(const char *tag);
-		static char getHtmlEntityLatin1Char(const char *entity);
-		static Common::String substituteHtmlEntities(const char *text);
-
-		void parseTagAttributes(const char* tagAttributes);
-		void cleanTextNode(int animDataSize);
-
-		HtmlTagType _currentTagType;
-		Common::SeekableReadStream *_stream;
-		int _pos;
-		int _posBak;
-		bool _field_10;
-		char _buffer[256];
-		HtmlMark _marks[kMaxNbrOfHtmlMarks];
-		uint32 _currentMarkIndex;
-		Common::String _htmlVariables[10];
-	};
-
 	INIConfig _inis;
 	TranslationDatabases _translationDatabases;
 	Common::HashMap<Common::String, Database, Common::IgnoreCase_Hash, Common::IgnoreCase_EqualTo> _databases;
